double fix

+++ PirateInterpreter/Values/FloatValue.cs
@@ -8,12 +8,12 @@ public class FloatValue : BaseValue, IValue
 
     public override BaseValue OperatedBy(Token _operator, BaseValue other)
     {
-        if (Value is not float && other.Value is not float)
+        if (Value is not double && other.Value is not double)
         {
-            throw new TypeConversionException(typeof(float));
+            throw new TypeConversionException(typeof(double));
         }
-        var value = (float)Value;
-        var otherValue = (float)other.Value;
+        var value = (double)Value;
+        var otherValue = (double)other.Value;
         switch (_operator.TokenType)
         {
             case TokenType.PLUS:
